+++ CHANGELOG.md
@@ -15,6 +15,7 @@ AdminNeo 5.2.0
 
 ### Bugfixes
 
+- MySQL: Fix retrieving additional info about the last query
 - MariaDB: Fix changing user password (issue #155)
 - MS SQL: Fix collation issues when retrieving default values (by @vrana)
 - PDO: Fix displaying the number of affected rows (regression from 5.1.0) (issue #151)